requirements.txt
playwright==1.40.0
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
python-dotenv==1.0.0
email-validator==2.0.0
python-multipart==0.0.6
jinja2==3.1.2
